requirements.txt
Requests==2.32.3
feedparser==6.0.11
python-dotenv==1.1.1
